requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
Pillow>=10.3.0
